fix(jobs): move runner flags outside shared
