+++ packages/jobs/lib/execution/onEvent.ts
@@ -2,11 +2,12 @@ import { Err, metrics, Ok } from '@nangohq/utils';
 import type { Result } from '@nangohq/utils';
 import type { TaskOnEvent } from '@nangohq/nango-orchestrator';
 import type { Config, SyncConfig, NangoConnection, NangoProps } from '@nangohq/shared';
-import { configService, environmentService, getApiUrl, getRunnerFlags, NangoError } from '@nangohq/shared';
+import { configService, environmentService, featureFlags, getApiUrl, NangoError } from '@nangohq/shared';
 import { logContextGetter } from '@nangohq/logs';
 import type { DBEnvironment, DBTeam } from '@nangohq/types';
 import { startScript } from './operations/start.js';
 import { bigQueryClient } from '../clients.js';
+import { getRunnerFlags } from '../utils/flags.js';
 
 export async function startOnEvent(task: TaskOnEvent): Promise<Result<void>> {
     let account: DBTeam | undefined;
@@ -70,7 +71,7 @@ export async function startOnEvent(task: TaskOnEvent): Promise<Result<void>> {
             nangoConnectionId: task.connection.id,
             syncConfig: syncConfig,
             debug: false,
-            runnerFlags: await getRunnerFlags(),
+            runnerFlags: await getRunnerFlags(featureFlags),
             startedAt: new Date()
         };
