+++ packages/jobs/lib/execution/sync.ts
@@ -5,7 +5,6 @@ import {
     externalWebhookService,
     getApiUrl,
     getLastSyncDate,
-    getRunnerFlags,
     updateSyncJobStatus,
     SyncStatus,
     errorManager,
@@ -21,7 +20,8 @@ import {
     configService,
     createSyncJob,
     getSyncConfigRaw,
-    getSyncJobByRunId
+    getSyncJobByRunId,
+    featureFlags
 } from '@nangohq/shared';
 import { Err, Ok, metrics } from '@nangohq/utils';
 import type { Result } from '@nangohq/utils';
@@ -35,6 +35,7 @@ import { records } from '@nangohq/records';
 import type { TaskSync, TaskSyncAbort } from '@nangohq/nango-orchestrator';
 import { abortScript } from './operations/abort.js';
 import { logger } from '../logger.js';
+import { getRunnerFlags } from '../utils/flags.js';
 
 export async function startSync(task: TaskSync, startScriptFn = startScript): Promise<Result<NangoProps>> {
     let logCtx: LogContext | undefined;
@@ -130,7 +131,7 @@ export async function startSync(task: TaskSync, startScriptFn = startScript): Pr
             track_deletes: syncConfig.track_deletes,
             syncConfig: syncConfig,
             debug: task.debug || false,
-            runnerFlags: await getRunnerFlags(),
+            runnerFlags: await getRunnerFlags(featureFlags),
             startedAt: new Date(),
             ...(lastSyncDate ? { lastSyncDate } : {})
         };
