+++ packages/shared/lib/index.ts
@@ -17,14 +17,13 @@ import syncManager, { syncCommandToOperation } from './services/sync/manager.ser
 import flowService from './services/flow.service.js';
 import { errorNotificationService } from './services/notification/error.service.js';
 import analytics, { AnalyticsTypes } from './utils/analytics.js';
-import featureFlags from './utils/featureflags.js';
+import featureFlags, { FeatureFlags } from './utils/featureflags.js';
 import { Orchestrator } from './clients/orchestrator.js';
 import { SlackService, generateSlackConnectionId } from './services/notification/slack.service.js';
 
 export * from './services/on-event-scripts.service.js';
 export * from './services/sync/sync.service.js';
 export * from './services/sync/job.service.js';
-export * from './services/sync/run.utils.js';
 export * from './services/sync/config/config.service.js';
 export * from './services/sync/config/endpoint.service.js';
 export * from './services/sync/config/deploy.service.js';
@@ -79,6 +78,7 @@ export {
     errorNotificationService,
     analytics,
     AnalyticsTypes,
+    FeatureFlags,
     featureFlags,
     syncCommandToOperation,
     Orchestrator,
